single patient info part implemented. Policies introduced to Patient class

+++ database/migrations/2016_03_24_162804_users_clinics_foreign.php
@@ -13,7 +13,7 @@ class UsersClinicsForeign extends Migration
     public function up()
     {
         Schema::table('users',function(Blueprint $table){
-            $table->integer('clinic_id')->unsigned();
+            $table->bigInteger('clinic_id')->unsigned();
 
             $table->foreign('clinic_id')->references('id')->on('clinics')->onDelete('restrict');
         });
